requirements.txt
aiogram
APScheduler
asyncio
asyncpg-lite
Cython
pandas
python-decouple
pipdeptree
requests
Telethon